+++ src/index.ts
@@ -148,6 +148,9 @@ export class TyntecWhatsAppAdapter extends BotAdapter {
             if (activity.attachments !== undefined) {
                 throw Error(`TyntecWhatsAppAdapter: both text Activity.channelData.contentType and Activity.attachments not supported: ${activity.channelData.contentType} and ${JSON.stringify(activity.attachments)}`);
             }
+            if (activity.channelData!.contacts !== undefined) {
+                throw Error(`TyntecWhatsAppAdapter: both text Activity.channelData.contentType and Activity.channelData.contacts not supported: ${activity.channelData.contentType} and ${JSON.stringify(activity.channelData.contacts)}`);
+            }
             if (activity.channelData!.location !== undefined) {
                 throw Error(`TyntecWhatsAppAdapter: both text Activity.channelData.contentType and Activity.channelData.location not supported: ${activity.channelData.contentType} and ${JSON.stringify(activity.channelData.location)}`);
             }
@@ -180,6 +183,9 @@ export class TyntecWhatsAppAdapter extends BotAdapter {
             }
             if (activity.attachments[0].thumbnailUrl !== undefined) {
                 console.warn(`TyntecWhatsAppAdapter: Activity.attachments.thumbnailUrl not supported: ${activity.attachments[0].thumbnailUrl}`);
+            }
+            if (activity.channelData!.contacts !== undefined) {
+                throw Error(`TyntecWhatsAppAdapter: both media Activity.channelData.contentType and Activity.channelData.contacts not supported: ${activity.channelData.contentType} and ${JSON.stringify(activity.channelData.contacts)}`);
             }
             if (activity.channelData!.location !== undefined) {
                 throw Error(`TyntecWhatsAppAdapter: both media Activity.channelData.contentType and Activity.channelData.location not supported: ${activity.channelData.contentType} and ${JSON.stringify(activity.channelData.location)}`);
@@ -266,9 +272,39 @@ export class TyntecWhatsAppAdapter extends BotAdapter {
                 };
             }
         }
+        if (activity.channelData!.contentType === "contacts") {
+            if (activity.attachments !== undefined) {
+                throw Error(`TyntecWhatsAppAdapter: both contacts Activity.channelData.contentType and Activity.attachments not supported: ${activity.channelData.contentType} and ${JSON.stringify(activity.attachments)}`);
+            }
+            if (activity.channelData!.contacts === undefined) {
+                throw Error(`TyntecWhatsAppAdapter: contacts Activity.channelData.contentType requires Activity.channelData.contacts: ${activity.channelData.contacts}`);
+            }
+            if (activity.channelData!.location !== undefined) {
+                throw Error(`TyntecWhatsAppAdapter: both contacts Activity.channelData.contentType and Activity.channelData.location not supported: ${activity.channelData.contentType} and ${JSON.stringify(activity.channelData.location)}`);
+            }
+            if (activity.channelData!.template !== undefined) {
+                throw Error(`TyntecWhatsAppAdapter: both contacts Activity.channelData.contentType and Activity.channelData.template not supported: ${activity.channelData.contentType} and ${JSON.stringify(activity.channelData.template)}`);
+            }
+            if (activity.text !== undefined) {
+                throw Error(`TyntecWhatsAppAdapter: both contacts Activity.channelData.contentType and Activity.text not supported: ${activity.channelData.contentType} and ${activity.text}`);
+            }
+
+            return {
+                from: activity.from.id,
+                to: activity.conversation.id,
+                channel: "whatsapp",
+                content: {
+                    contentType: "contacts",
+                    contacts: activity.channelData!.contacts
+                }
+            };
+        }
         if (activity.channelData!.contentType === "location") {
             if (activity.attachments !== undefined) {
                 throw Error(`TyntecWhatsAppAdapter: both location Activity.channelData.contentType and Activity.attachments not supported: ${activity.channelData.contentType} and ${JSON.stringify(activity.attachments)}`);
+            }
+            if (activity.channelData!.contacts !== undefined) {
+                throw Error(`TyntecWhatsAppAdapter: both location Activity.channelData.contentType and Activity.channelData.contacts not supported: ${activity.channelData.contentType} and ${JSON.stringify(activity.channelData.contacts)}`);
             }
             if (activity.channelData!.location === undefined) {
                 throw Error(`TyntecWhatsAppAdapter: location Activity.channelData.contentType requires Activity.channelData.location: ${activity.channelData.location}`);
@@ -293,6 +329,9 @@ export class TyntecWhatsAppAdapter extends BotAdapter {
         if (activity.channelData!.contentType === "template") {
             if (activity.attachments !== undefined) {
                 throw Error(`TyntecWhatsAppAdapter: both Activity.attachments and template Activity.channelData.contentType not supported: ${JSON.stringify(activity.attachments)} and ${activity.channelData.contentType}`);
+            }
+            if (activity.channelData!.contacts !== undefined) {
+                throw Error(`TyntecWhatsAppAdapter: both Activity.channelData.contacts and template Activity.channelData.contentType not supported: ${JSON.stringify(activity.channelData.contacts)} and ${activity.channelData.contentType}`);
             }
             if (activity.channelData!.location !== undefined) {
                 throw Error(`TyntecWhatsAppAdapter: both Activity.channelData.location and template Activity.channelData.contentType not supported: ${JSON.stringify(activity.channelData.location)} and ${activity.channelData.contentType}`);