+++ CHANGELOG.md
@@ -8,4 +8,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Support for processing contacts messages to `TyntecWhatsAppAdapter#processActivity`.
 - Support for processing location messages to `TyntecWhatsAppAdapter#processActivity`.
+- Support for sending contacts messages to `TyntecWhatsAppAdapter#sendActivities`.
 - Support for sending location messages to `TyntecWhatsAppAdapter#sendActivities`.